requirements.txt
crewai
crewai[tools]
langchain-openai
langchain-community
python-dotenv